fix(lti): open the deep-link picker in a severed independent tab (escape CAA)
Break-out attempt #1 (window.open handoff window) still failed: a popup that
keeps a window.opener link to the Schoology iframe inherits the iframe's
storage partition, so Google Workspace Context-Aware Access still denies the
sign-in token ("Account Restricted"). The only context that passes CAA is a
fully independent, opener-less top-level tab (confirmed: the teacher's own
spartboard.web.app tab signs in fine).

Rebuild the picker as two modes dispatched on isFramed():
- Framed (in Schoology): a thin LAUNCHER whose "Choose a quiz" opens this same
  launch URL in a severed tab via a click-driven <a target=_blank
  rel="noopener noreferrer">. It does NOT consume the one-time launch code.
- Not framed (that tab): the full FLOW — exchange the launch code, Google
  sign-in (now first-party → CAA passes), load library, create assignment, sign
  the LtiDeepLinkingResponse, and form-POST it back to Schoology from the tab.

Severing the opener also severs every cross-window channel, so the tab finishes
the deep-link return itself (the Add-Materials modal can't auto-close; the
teacher closes the tab and refreshes course materials). Replaces the
postMessage handoff (deepLinkHandoff.ts + LtiDeepLinkWindow.tsx removed). The
return-URL https + schoology.com validation (CodeQL fix) is preserved inline.
